requirements.txt
asgiref==3.3.4
cachetools==4.2.4
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
charset-normalizer==2.0.3
cloudinary==1.26.0
cryptography==3.4.7
defusedxml==0.7.1
dj-database-url==0.5.0
dj-rest-auth==2.1.7
dj3-cloudinary-storage==0.0.6
Django==3.2.4
django-allauth==0.44.0
django-cloudinary-storage==0.2.3
django-cors-headers==3.7.0
django-environ==0.4.5
django-heroku==0.3.1
django-rest-knox==4.1.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
env==0.1.0
google-api-core==2.3.2
google-api-python-client==2.33.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
googleapis-common-protos==1.54.0
gunicorn==20.1.0
haversine==2.3.1
httplib2==0.20.2
idna==2.10
mock==4.0.3
oauthlib==3.1.1
Pillow==8.2.0
protobuf==3.19.1
psycopg2==2.9.1
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==2.1.0
pyparsing==3.0.6
python-environ==0.4.54
python3-openid==3.2.0
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.8
six==1.16.0
sqlparse==0.4.1
typing-extensions==3.10.0.0
uritemplate==4.1.1
urllib3==1.26.5
whitenoise==5.3.0
python-dateutil==2.8.1